% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender.R
\docType{data}
\name{df}
\alias{df}
\title{French first names 1900-2016}
\format{A data frame with 620994 rows and 4 variables: \describe{
  \item{sex}{Sex, with two levels, 1 or 2}
  \item{firstname}{First name}
  \item{year}{Year of birth}
  \item{count}{Number of individuals born that year}
  }}
\source{
\url{https://www.insee.fr/fr/statistiques/2540004}
}
\usage{
df
}
\description{
A dataset containing the first name, sex and count of individuals born in
France for each year between 2000 and 2016 attributes of almost 54,000
}
\keyword{datasets}
