% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender.R
\name{gender}
\alias{gender}
\title{Predict genders from first names}
\usage{
gender(firstname, freq = FALSE)
}
\arguments{
\item{firstname}{First names as a character vector.}

\item{freq}{Return the probability of the first names being males.}
}
\value{
The predicted genders based on the proportions of males and females with the input first names. Possible values are either "male" or "female". NAs are returned when the input first names are unknown in the database.
If freq is set to TRUE, the function returns the probabilities of the first names being male.
}
\description{
This function predicts the genders of a vector of first names.
}
\examples{
gender(c("Baptiste", "Henriette")
}
\seealso{
\code{\link{gender_unique}}, \code{\link{is.male}}, \code{\link{is.female}}
}
